split and reallign - no changes
